pyproject.toml
[project]
name = "cloneofx"
version = "0.1.0"
description = ""
authors = [
    {name = "Alison Oliveira",email = "alison.oliveira89@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.2.2,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "gitpython (>=3.1.44,<4.0.0)" 
]

[tool.poetry]
packages = [{include = "cloneofx"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
